+++ composeApp/src/jvmMain/kotlin/org/malv/descuentados/Main.kt
@@ -1,4 +1,4 @@
-package org.malv.descontados
+package org.malv.descuentados
 
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.unit.dp
@@ -6,8 +6,8 @@ import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.WindowPosition
 import androidx.compose.ui.window.WindowState
 import androidx.compose.ui.window.application
-import org.malv.descontados.services.ConfigurationService
-import org.malv.descontados.ui.App
+import org.malv.descuentados.services.ConfigurationService
+import org.malv.descuentados.ui.App
 
 fun main(args: Array<String>) = application {
     Window(
@@ -15,7 +15,7 @@ fun main(args: Array<String>) = application {
             ConfigurationService.instance.save()
             exitApplication()
         },
-        title = "Descontados",
+        title = "Descuentados",
         state = WindowState(width = 1000.dp, height = 800.dp, position = WindowPosition.Aligned(Alignment.TopEnd))
     ) {
         App()
