+++ composeApp/src/jvmMain/kotlin/org/malv/descuentados/models/VideoResult.kt
@@ -1,4 +1,4 @@
-package org.malv.descontados.models
+package org.malv.descuentados.models
 
 data class VideoResult(
     val videoId: String,
