+++ composeApp/src/jvmMain/kotlin/org/malv/descuentados/services/LoginService.kt
@@ -1,4 +1,4 @@
-package org.malv.descontados.services
+package org.malv.descuentados.services
 
 import com.google.api.client.auth.oauth2.Credential
 import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp
