+++ composeApp/src/jvmMain/kotlin/org/malv/descuentados/ui/AppComponents.kt
@@ -1,4 +1,4 @@
-package org.malv.descontados.ui
+package org.malv.descuentados.ui
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -19,7 +19,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.rememberNavController
-import org.malv.descontados.models.Routes
+import org.malv.descuentados.models.Routes
 
 @Composable
 @OptIn(ExperimentalMaterial3Api::class)
