+++ composeApp/src/jvmMain/kotlin/org/malv/descuentados/ui/SettingsComponents.kt
@@ -1,4 +1,4 @@
-package org.malv.descontados.ui
+package org.malv.descuentados.ui
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -20,9 +20,9 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import org.malv.descontados.models.Language
-import org.malv.descontados.services.ConfigurationService
-import org.malv.descontados.viewmodels.CodesViewModel
+import org.malv.descuentados.models.Language
+import org.malv.descuentados.services.ConfigurationService
+import org.malv.descuentados.viewmodels.CodesViewModel
 
 @Composable
 fun SettingsUI(
